requirements.txt
blinker==1.9.0
click==8.1.8
flask==3.1.0
Flask-Cors==5.0.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing-extensions==4.12.2
werkzeug==3.1.3
zipp==3.21.0
